Expose custom theme functionality
This commit also reworks the `themes` example to utilize it, which
unfortunately means its blue example theme is gone for now. If a formal
mechanism to switch themes while the window is open is implemented, the
blue theme can be reintroduced.
