+++ examples/custom-theme/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "themes"
+name = "custom-theme"
 version = "0.1.0"
 edition = "2021"
